+++ smmregrid/cdogenerate.py
@@ -214,17 +214,15 @@ class CdoGenerate():
                               })
                 ppp.start()
                 processes.append(ppp)
-
             for proc in processes:
                 proc.join()
 
-        weights = self.weightslist_to_3d(wlist, method, mask_dim)
+        weights = self.weightslist_to_3d(wlist, method, sgrid[mask_dim])
         weights_matrix = compute_weights_matrix3d(weights, mask_dim)
         weights = mask_weights(weights, weights_matrix, mask_dim)
         masked = check_mask(weights, mask_dim)
-        masked = [int(x) for x in masked]
         masked_xa = xarray.DataArray(masked,
-                                     coords={mask_dim: range(0, len(masked))},
+                                     coords={mask_dim: weights[mask_dim]},
                                      name="dst_grid_masked")
 
         return xarray.merge([weights, masked_xa])
@@ -309,32 +307,39 @@ class CdoGenerate():
     #     if not isinstance(tmpfile, str):
     #         os.remove(tmpfile.name)
 
-    def weightslist_to_3d(self, ds_list, method='ycon', mask_dim='lev'):
-        """Combine a list of 2D CDO weights into a 3D one."""
+    def weightslist_to_3d(self, weights_list, method='ycon', mask_coord=xarray.DataArray):
+        """Combine a list of 2D CDO weights into a 3D one.
+        
+        Args:
+            ds_list (list): List of xarray.Dataset with 2D weights
+            method (str): Remap method, to determine if dst_grid_area and dst_grid_frac
+                          need to be included
+            mask_coord (xarray.DataArray): Coordinate to use for the 3D weights
+        """
 
-        links_dim = "numLinks" if "numLinks" in ds_list[0].dims else "num_links"
-        dim_values = range(len(ds_list))
-        nl = [ds.src_address.size for ds in ds_list]
-        nl0 = max(nl)
-        nlda = xarray.DataArray(nl, coords={mask_dim: range(0, len(nl))}, name="link_length")
+        links_dim = "numLinks" if "numLinks" in weights_list[0].dims else "num_links"
+        number_links = [ds.src_address.size for ds in weights_list]
+        links_array = xarray.DataArray(number_links, coords={mask_coord.name: mask_coord.values}, name="link_length")
 
-        new_array = []
         varlist = ["src_address", "dst_address", "remap_matrix", "src_grid_imask", "dst_grid_imask"]
         # Add dst_grid_area and dst_grid_frac only for conservative methods
         if method in ['ycon', 'con2', 'con']:
             varlist += ["dst_grid_area", "dst_grid_frac"]
-        ds0 = ds_list[0].drop_vars(varlist)
+        untouched_array = weights_list[0].drop_vars(varlist)
 
-        for x, d in zip(ds_list, dim_values):
+        modified_array  = []
+        nl_max = max(number_links)
+        for x, v in zip(weights_list, mask_coord.values):
             nl1 = x.src_address.size
-            xplist = [x[vname].pad(**{links_dim: (0, nl0 - nl1), "mode": 'constant', "constant_values": 0})
+            xplist = [x[vname].pad(**{links_dim: (0, nl_max - nl1), "mode": 'constant', "constant_values": 0})
                       for vname in varlist]
             xmerged = xarray.merge(xplist)
-            new_array.append(xmerged.assign_coords({mask_dim: d}))
+            modified_array.append(xmerged.assign_coords({mask_coord.name: v}))
+        modified_array =  xarray.concat(modified_array, mask_coord.name, coords='different', compat='equals')
 
-        return xarray.merge([nlda, ds0, xarray.concat(new_array, mask_dim, coords='different', compat='equals')],
+        return xarray.merge([links_array, untouched_array, modified_array],
                             combine_attrs='no_conflicts')
-
+    
     def areas(self, target=False):
         """Generate source areas or target areas"""