+++ tests/levels_test.py
@@ -12,17 +12,17 @@ tfile = os.path.join(INDIR, 'r360x180.nc')
 
 @pytest.mark.parametrize("method", ['con'])
 def test_plev_gaussian_levels(method):
+    """Pressure levels on gaussian grid test with level selection"""
     fff = check_cdo_regrid_levels(os.path.join(INDIR, 'ua-ecearth.nc'), tfile,
                                   "plev", [14, 15, 17], remap_method=method)
     assert fff is True
 
-# 3D NEMO grid test
-
 
 @pytest.mark.parametrize("method", ['nn'])
 def test_nemo_3d_levels(method):
+    """NEMO 3D levels test with level selection"""
     fff = check_cdo_regrid_levels(os.path.join(INDIR, 'so3d-nemo.nc'), tfile,
-                                  "lev", [14, 15, 17], remap_method=method)
+                                  "lev", [14, 17], remap_method=method)
     assert fff is True, "Multiple levels test failed"
 
     fff = check_cdo_regrid_levels(os.path.join(INDIR, 'so3d-nemo.nc'), tfile,