+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+- Introduction `check_nan` option to detect if a vertical dimension has NaN values (#50)
+
 ## [v0.1.3]
 
 - Improve grid kind and bounds detection via `GridInspector` (#48, #49)