+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+- Replace `vertical_dim` to more correct `mask_dim` as changing direction for mask (#56)
 - Introduction `check_nan` option to detect if a vertical dimension has NaN values (#50)
 
 ## [v0.1.3]